+++ src/internal/connector/exchange/data_collections.go
@@ -170,7 +170,6 @@ func DataCollections(
 	acct account.M365Config,
 	su support.StatusUpdater,
 	ctrlOpts control.Options,
-	canMakeDeltaQueries bool,
 	errs *fault.Bus,
 ) ([]data.BackupCollection, *prefixmatcher.StringSetMatcher, error) {
 	eb, err := selector.ToExchangeBackup()
@@ -207,7 +206,6 @@ func DataCollections(
 			scope,
 			cdps[scope.Category().PathType()],
 			ctrlOpts,
-			canMakeDeltaQueries,
 			su,
 			errs)
 		if err != nil {
@@ -263,7 +261,6 @@ func createCollections(
 	scope selectors.ExchangeScope,
 	dps DeltaPaths,
 	ctrlOpts control.Options,
-	canMakeDeltaQueries bool,
 	su support.StatusUpdater,
 	errs *fault.Bus,
 ) ([]data.BackupCollection, error) {
@@ -306,7 +303,6 @@ func createCollections(
 		scope,
 		dps,
 		ctrlOpts,
-		canMakeDeltaQueries,
 		errs,
 	)
 	if err != nil {