+++ src/internal/connector/exchange/service_iterators.go
@@ -45,7 +45,6 @@ func filterContainersAndFillCollections(
 	scope selectors.ExchangeScope,
 	dps DeltaPaths,
 	ctrlOpts control.Options,
-	canMakeDeltaQueries bool,
 	errs *fault.Bus,
 ) (map[string]data.BackupCollection, error) {
 	var (
@@ -173,7 +172,7 @@ func filterContainersAndFillCollections(
 			cID,
 			prevDelta,
 			ctrlOpts.ToggleFeatures.ExchangeImmutableIDs,
-			canMakeDeltaQueries,
+			!ctrlOpts.ToggleFeatures.DisableDelta,
 		)
 		if err != nil {
 			if !graph.IsErrDeletedInFlight(err) {